+++ packages/core-ethereum/src/nonce-tracker.spec.ts
@@ -278,7 +278,7 @@ describe('nonce-tracker', function () {
 
 const genTx = (opts: { nonce: number; createdAt?: number }): Transaction => {
   const { createdAt = new Date().getTime() } = opts
-  return { ...opts, from: USER_ADDRESS.toHex(), createdAt, maxPrority: MAX_PRIORITY_FEE }
+  return { ...opts, from: USER_ADDRESS.toHex(), createdAt, maxPriority: MAX_PRIORITY_FEE, maxFeePerGas: BigNumber.from(10e9), gasLimit: BigNumber.from(400e3) }
 }
 
 const genMultiTx = (opts: {