+++ packages/core-ethereum/src/transaction-manager.spec.ts
@@ -2,7 +2,12 @@ import { expect } from 'chai'
 import { BigNumber } from 'ethers'
 import TransactionManager, { Transaction, TransactionPayload } from './transaction-manager.js'
 
-const TX: [string, Omit<Transaction, 'createdAt'>] = ['0', { nonce: 0, maxPrority: BigNumber.from('2000000000') }]
+const TX: [string, Omit<Transaction, 'createdAt'>] = ['0', { 
+  nonce: 0, 
+  maxPriority: BigNumber.from('2000000000'),
+  maxFeePerGas: BigNumber.from(10e9),
+  gasLimit: BigNumber.from(400e3)
+}]
 const PAYLOAD: TransactionPayload = { to: '0x0', data: '0x123', value: BigNumber.from('1') }
 
 describe('transaction-manager', function () {
@@ -94,7 +99,7 @@ describe('transaction-manager', function () {
 
     // generate mock txs
     for (let i = 0; i < 7; i++) {
-      txs.push([String(i), { nonce: i, maxPrority: BigNumber.from('1') }])
+      txs.push([String(i), { nonce: i, maxPriority: BigNumber.from('1'), maxFeePerGas: BigNumber.from(10), gasLimit: BigNumber.from(400e3)}])
     }
 
     // add them to confirmed
@@ -124,7 +129,9 @@ describe('transaction-manager', function () {
       data: PAYLOAD.data,
       value: PAYLOAD.value,
       nonce: TX[1].nonce,
-      maxPrority: TX[1].maxPrority
+      maxPriorityFeePerGas: TX[1].maxPriority,
+      maxFeePerGas: TX[1].maxFeePerGas,
+      gasLimit: TX[1].gasLimit
     })
   })
 })